requirements.txt
customtkinter
python-magic
SpeechRecognition
scikit-learn
numpy
pdfplumber
python-pptx
pefile
pyttsx3
Pillow
gtts
nmap
scikit-learn
